+++ packages/dx-react-scheduler-material-ui/src/templates/appointment-tooltip/content.test.jsx
@@ -105,7 +105,7 @@ describe('Appointment Tooltip', () => {
       expect(viewBoundText)
         .toHaveBeenCalledWith(
           defaultProps.appointmentData.startDate,
-          defaultProps.appointmentData.endDate, '',
+          defaultProps.appointmentData.endDate, 'appointmentTooltip',
           defaultProps.appointmentData.startDate, 1,
           defaultProps.formatDate,
         );
