+++ packages/dx-scheduler-core/src/constants.ts
@@ -34,6 +34,7 @@ export const DAY_SHORT_MONTH_SHORT_YEAR_OPTIONS = {
 export const LONG_WEEK_DAY_OPTIONS = { weekday: 'long' };
 export const LONG_MONTH_OPTIONS = { month: 'long' };
 export const DAY_LONG_MONTH_OPTIONS = { day: 'numeric', month: 'long' };
+export const EMPTY_OPTIONS = {};
 
 export const RECURRENCE_EDIT_SCOPE = {
   ALL: 'all',
@@ -43,3 +44,5 @@ export const RECURRENCE_EDIT_SCOPE = {
 
 export const TOGGLE_APPOINTMENT_FORM_VISIBILITY = 'toggleAppointmentFormVisibility';
 export const TOGGLE_APPOINTMENT_TOOLTIP_VISIBILITY = 'toggleAppointmentTooltipVisibility';
+
+export const APPOINTMENT_TOOLTIP = 'appointmentTooltip';
